feat: impl storage service with proxy
